# gh-117607: Speedup `os.path.relpath()`