--- a/Lib/posixpath.py
+++ b/Lib/posixpath.py
@@ -510,12 +510,10 @@ def relpath(path, start=None):
         curdir = b'.'
         sep = b'/'
         pardir = b'..'
-        getcwd = os.getcwdb
     else:
         curdir = '.'
         sep = '/'
         pardir = '..'
-        getcwd = os.getcwd
 
     if start is None:
         start = curdir
@@ -523,10 +521,8 @@ def relpath(path, start=None):
         start = os.fspath(start)
 
     try:
-        start_abs = getcwd() if not start or start == curdir else abspath(start)
-        path_abs = abspath(path)
-        start_tail = start_abs.lstrip(sep)
-        path_tail = path_abs.lstrip(sep)
+        start_tail = abspath(start).lstrip(sep)
+        path_tail = abspath(path).lstrip(sep)
         start_list = start_tail.split(sep) if start_tail else []
         path_list = path_tail.split(sep) if path_tail else []
         # Work out how much of the filepath is shared by start and path.