--- a/Misc/NEWS.d/next/Core and Builtins/2024-04-07-18-42-09.gh-issue-117607.C978BD.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2024-04-07-18-42-09.gh-issue-117607.C978BD.rst
@@ -1,1 +1,1 @@
-Speedup :func:`os.path.relpath()`.
+Speedup :func:`os.path.relpath`.