Benchmark:

- `ntpath.py`
    ```bat
    ::test.bat
    @echo off
    echo relpath() && python -m timeit -s "import before.ntpath" "before.ntpath.relpath('foo')" && python -m timeit -s "import after.ntpath" "after.ntpath.relpath('foo')"
    echo relpath('C:/') && python -m timeit -s "import before.ntpath" "before.ntpath.relpath('C:/')" && python -m timeit -s "import after.ntpath" "after.ntpath.relpath('C:/')"
    ```

    ```none
    relpath() 
    50000 loops, best of 5: 8.95 usec per loop # before
    50000 loops, best of 5: 7.11 usec per loop # after
    # -> 1.26x faster
    relpath('C:/')
    50000 loops, best of 5: 7 usec per loop # before
    100000 loops, best of 5: 3.61 usec per loop # after
    # -> 1.94x faster (for absolute paths)
    ```
- `posixpath.py`
    **TODO**