@nineteendo 2024-04-17
@eryksun, could you run a benchmark for `os.path.relpath()` on your machine? Otherwise it might undersell the performance improvement.