@nineteendo 2024-04-17
Could anyone do a benchmark of Unix? `os.getcwd()` takes way too long on my machine.