@eryksun 2024-04-17
> @eryksun, could you run a benchmark for `os.path.relpath()` on your machine? Otherwise it might undersell the performance improvement.

Windows - ntpath - 1.407-1.565 speedup

```text
> .\python_main -m timeit -s "import ntpath" "ntpath.relpath('C:foo', start='C:/Users/account')"
50000 loops, best of 5: 6.7 usec per loop
> .\python_work -m timeit -s "import ntpath" "ntpath.relpath('C:foo', start='C:/Users/account')"
50000 loops, best of 5: 4.28 usec per loop

> .\python_main -m timeit -s "import ntpath" "ntpath.relpath('C:/', start='C:/Users/account')"
50000 loops, best of 5: 4.84 usec per loop
> .\python_work -m timeit -s "import ntpath" "ntpath.relpath('C:/', start='C:/Users/account')"
100000 loops, best of 5: 3.44 usec per loop
```

Linux - posixpath - 1.137-1.189 speedup

```text
$ ./python_main -m timeit -s "import posixpath" "posixpath.relpath('foo', start='/home/account')"
100000 loops, best of 5: 3.14 usec per loop
$ ./python_work -m timeit -s "import posixpath" "posixpath.relpath('foo', start='/home/account')"
100000 loops, best of 5: 2.64 usec per loop

$ ./python_main -m timeit -s "import posixpath" "posixpath.relpath('/', start='/home/account')"
200000 loops, best of 5: 1.74 usec per loop
$ ./python_work -m timeit -s "import posixpath" "posixpath.relpath('/', start='/home/account')"
200000 loops, best of 5: 1.53 usec per loop
```

Windows - posixpath - 1.123-1.136 speedup

```text
> .\python_main -m timeit -s "import posixpath" "posixpath.relpath('foo', start='/home/account')"
50000 loops, best of 5: 4.39 usec per loop
> .\python_work -m timeit -s "import posixpath" "posixpath.relpath('foo', start='/home/account')"
100000 loops, best of 5: 3.91 usec per loop

> .\python_main -m timeit -s "import posixpath" "posixpath.relpath('/', start='/home/account')"
100000 loops, best of 5: 3.1 usec per loop
> .\python_work -m timeit -s "import posixpath" "posixpath.relpath('/', start='/home/account')"
100000 loops, best of 5: 2.73 usec per loop
```

Linux - ntpath - 1.006-1.007 slowdown (no change)

```text
$ ./python_main -m timeit -s "import ntpath" "ntpath.relpath('C:foo', start='C:/Users/account')"
50000 loops, best of 5: 8.63 usec per loop
$ ./python_work -m timeit -s "import ntpath" "ntpath.relpath('C:foo', start='C:/Users/account')"
50000 loops, best of 5: 8.69 usec per loop

$ ./python_main -m timeit -s "import ntpath" "ntpath.relpath('C:/', start='C:/Users/account')"
50000 loops, best of 5: 6.29 usec per loop
$ ./python_work -m timeit -s "import ntpath" "ntpath.relpath('C:/', start='C:/Users/account')"
50000 loops, best of 5: 6.33 usec per loop
```