@eryksun 2024-04-17
> Could you also check with `start` not set or a relative path?

Sure, except testing `ntpath.relpath("C:/")` isn't meaningful on POSIX. It fails because the "C:" drive is a different device namespace.

Windows - ntpath - 1.201-1.269 speedup

```text
> .\python_main -m timeit -s "import os,ntpath; os.chdir('C:/Users')" "ntpath.relpath('foo')"
50000 loops, best of 5: 4.84 usec per loop
> .\python_work -m timeit -s "import os,ntpath; os.chdir('C:/Users')" "ntpath.relpath('foo')"
50000 loops, best of 5: 4.03 usec per loop

> .\python_main -m timeit -s "import os,ntpath; os.chdir('C:/Users')" "ntpath.relpath('C:/')"
50000 loops, best of 5: 4.11 usec per loop
> .\python_work -m timeit -s "import os,ntpath; os.chdir('C:/Users')" "ntpath.relpath('C:/')"
100000 loops, best of 5: 3.24 usec per loop
```

Linux - posixpath - 1.309-1.434 speedup

```text
$ ./python_main -m timeit -s "import os,posixpath; os.chdir('/home')" "posixpath.relpath('foo')"
100000 loops, best of 5: 2.94 usec per loop
$ ./python_work -m timeit -s "import os,posixpath; os.chdir('/home')" "posixpath.relpath('foo')"
100000 loops, best of 5: 2.05 usec per loop

$ ./python_main -m timeit -s "import os,posixpath; os.chdir('/home')" "posixpath.relpath('/')"
100000 loops, best of 5: 2.12 usec per loop
$ ./python_work -m timeit -s "import os,posixpath; os.chdir('/home')" "posixpath.relpath('/')"
200000 loops, best of 5: 1.62 usec per loop
```

Windows - posixpath - 1.060 slowdown, 1.019 speedup

```text
> .\python_main -m timeit -s "import os,posixpath; os.chdir('C:/Users')" "posixpath.relpath('foo')"
50000 loops, best of 5: 4.01 usec per loop
> .\python_work -m timeit -s "import os,posixpath; os.chdir('C:/Users')" "posixpath.relpath('foo')"
50000 loops, best of 5: 4.25 usec per loop

> .\python_main -m timeit -s "import os,posixpath; os.chdir('C:/Users')" "posixpath.relpath('/')"
100000 loops, best of 5: 3.22 usec per loop
> .\python_work -m timeit -s "import os,posixpath; os.chdir('C:/Users')" "posixpath.relpath('/')"
100000 loops, best of 5: 3.16 usec per loop
```

Linux - ntpath - 1.004 speedup (no change)

```text
$ ./python_main -m timeit -s "import os,ntpath; os.chdir('/home')" "ntpath.relpath('foo')"
50000 loops, best of 5: 8.23 usec per loop
$ ./python_work -m timeit -s "import os,ntpath; os.chdir('/home')" "ntpath.relpath('foo')"
50000 loops, best of 5: 8.2 usec per loop
```